IOS-4456 Revert "IOS-4456: Add SendBlockedReason"
This reverts commit 4ea0ecabba55c2b83ce45f46e50aaf8e18824133.
