--- Tangem/App/ViewModels/WalletModel/WalletModel+State.swift
@@ -1,90 +1,0 @@
-//
-//  WalletModel+State.swift
-//  Tangem
-//
-//  Created by Andrew Son on 04/09/23.
-//  Copyright © 2023 Tangem AG. All rights reserved.
-//
-
-import Foundation
-
-extension WalletModel {
-    enum State: Hashable {
-        case created
-        case idle
-        case loading
-        case noAccount(message: String)
-        case failed(error: String)
-        case noDerivation
-
-        var isLoading: Bool {
-            switch self {
-            case .loading, .created:
-                return true
-            default:
-                return false
-            }
-        }
-
-        var isSuccesfullyLoaded: Bool {
-            switch self {
-            case .idle, .noAccount:
-                return true
-            default:
-                return false
-            }
-        }
-
-        var isBlockchainUnreachable: Bool {
-            switch self {
-            case .failed:
-                return true
-            default:
-                return false
-            }
-        }
-
-        var isNoAccount: Bool {
-            switch self {
-            case .noAccount:
-                return true
-            default:
-                return false
-            }
-        }
-
-        var errorDescription: String? {
-            switch self {
-            case .failed(let localizedDescription):
-                return localizedDescription
-            case .noAccount(let message):
-                return message
-            default:
-                return nil
-            }
-        }
-
-        var failureDescription: String? {
-            switch self {
-            case .failed(let localizedDescription):
-                return localizedDescription
-            default:
-                return nil
-            }
-        }
-
-        fileprivate var canCreateOrPurgeWallet: Bool {
-            switch self {
-            case .failed, .loading, .created, .noDerivation:
-                return false
-            case .noAccount, .idle:
-                return true
-            }
-        }
-    }
-
-    enum WalletManagerUpdateResult: Hashable {
-        case success
-        case noAccount(message: String)
-    }
-}
